move scalacheck Pipeline utils to separate trait
